Rectangles! Also I made rustdoc work a lot better. Next to add is drawing rectangles.

+++ src/engine/display.rs
@@ -1,7 +1,10 @@
 use super::{Engine, GameTrait, hal::HAL};
 
+/// Width of the display in pixels
 pub const DISPLAY_WIDTH: u8 = 128;
-pub const DISPLAY_HEIGHT: u8 = 64; // Must be a multiple of 8
+/// Height of the display in pixels, must be a multiple of 8
+pub const DISPLAY_HEIGHT: u8 = 64;
+/// Number of display pages
 pub const DISPLAY_PAGE_COUNT: u8 = DISPLAY_HEIGHT / 8;
 
 pub type Buffer = [[u8; DISPLAY_WIDTH as usize]; DISPLAY_PAGE_COUNT as usize];
@@ -22,32 +25,32 @@ impl FrameBuffer {
         hal.display_buffer(&self);
     }
 
-    pub(super) fn set_pixel_state(&mut self, x: usize, y: usize, state: bool) {
+    pub(super) fn set_pixel_state(&mut self, x: u32, y: u32, state: bool) {
         let page_no = Self::get_pixel_page_no(y);
         let y_coord_in_page = y % 8;
 
         let bitmask: u8 = 0x01 << y_coord_in_page;
         if state {
-            self.buffer[page_no][x] |= bitmask;
+            self.buffer[page_no as usize][x as usize] |= bitmask;
         } else {
-            self.buffer[page_no][x] &= !bitmask;
+            self.buffer[page_no as usize][x as usize] &= !bitmask;
         }
     }
 
-    pub(super) fn get_pixel_state(&self, x: usize, y: usize) -> bool {
+    pub(super) fn get_pixel_state(&self, x: u32, y: u32) -> bool {
         let page_no = Self::get_pixel_page_no(y);
         let y_coord_in_page = y % 8;
 
         let bitmask: u8 = 0x01 << y_coord_in_page;
 
-        (self.buffer[page_no][x] & bitmask) > 0
+        (self.buffer[page_no as usize][x as usize] & bitmask) > 0
     }
 
     pub(super) fn clear(&mut self) {
         self.buffer = [[0x00; DISPLAY_WIDTH as usize]; DISPLAY_PAGE_COUNT as usize];
     }
 
-    fn get_pixel_page_no(y: usize) -> usize {
+    fn get_pixel_page_no(y: u32) -> u32 {
         y / 8 // Each page is 8 pixels tall
     }
 }
